services/crypta/requirements.txt
asgiref==3.9.1
astroid==3.3.10
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.5
dill==0.4.0
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
et_xmlfile==2.0.0
idna==3.10
isort==6.0.1
jwt==1.4.0
mccabe==0.7.0
numpy==2.3.1
openpyxl==3.1.5
pandas==2.3.1
pillow==11.3.0
platformdirs==4.3.8
pycparser==2.22
pylint==3.3.7
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tomlkit==0.13.3
tzdata==2025.2
urllib3==2.5.0
